requirements.txt
apache-bookkeeper-client==4.13.0
certifi==2020.12.5
chardet==4.0.0
coloredlogs==15.0
fastavro==1.3.5
grpcio==1.27.2
humanfriendly==9.1
idna==2.10
prometheus-client==0.10.1
protobuf==3.15.8
pulsar-client==2.6.3
pymmh3==0.0.5
pytz==2021.1
ratelimit==2.2.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
